Adding all functions and services

+++ src/main/java/com/verdemar/verdemar/repository/ApartamentRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.verdemar.verdemar.domain.Apartament;
 
-public interface ApartamentRepository extends JpaRepository<Apartament, Long> {
+public interface ApartamentRepository extends JpaRepository<Apartament, Short> {
     // Custom query methods can be defined here if needed
     // For example, you can add methods to find apartments by specific criteria
     // or to perform complex queries using JPQL or native SQL.
